Move handle escape logic from modal overlay to modal component
